+++ docs/temml.cjs.js
@@ -151,7 +151,7 @@ const assert = function(value) {
   return value;
 };
 
-const textAtomTypes = ["text", "textord", "mathord"];
+const textAtomTypes = ["text", "textord", "mathord", "atom"];
 
 /**
  * Return the protocol of a URL, or "_relative" if the URL does not specify a
@@ -431,10 +431,6 @@ const toNode = function(tagName) {
 const toMarkup = function(tagName) {
   let markup = `<${tagName}`;
 
-  if (tagName === "svg") {
-    markup += ' xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink"';
-  }
-
   // Add the class
   if (this.classes.length) {
     markup += ` class="${utils.escape(createClass(this.classes))}"`;
@@ -478,8 +474,8 @@ const toMarkup = function(tagName) {
  *
  */
 class Span {
-  constructor(classes, children) {
-    initNode.call(this, classes);
+  constructor(classes, children, style) {
+    initNode.call(this, classes, style);
     this.children = children || [];
   }
 
@@ -732,7 +728,8 @@ var mathMLTree = {
  * This file provides support for building horizontal stretchy elements.
  */
 
-const keysWithoutUnicodePoints = ["longRightleftharpoons", "longLeftrightharpoons"];
+// From mhchem reaction arrows \ce{A <=>> B} and \ce{A <<=> B}
+const keysWithoutUnicodePoints = ["equilibriumRight", "equilibriumLeft"];
 
 const stretchyCodePoint = {
   widehat: "^",
@@ -768,92 +765,47 @@ const stretchyCodePoint = {
   xmapsto: "\u21a6",
   xrightharpoondown: "\u21c1",
   xleftharpoondown: "\u21bd",
-  xrightleftharpoons: "\u21cc",
-  xleftrightharpoons: "\u21cb",
   xtwoheadleftarrow: "\u219e",
   xtwoheadrightarrow: "\u21a0",
   xlongequal: "=",
-  xtofrom: "\u21c4",
   xrightleftarrows: "\u21c4",
-  longleftrightarrows: "\u21c4",
-  longrightleftharpoons: "\u21cc",
+  yields: "\u2192",
+  yieldsLeft: "\u2190",
+  mesomerism: "\u2194",
+  longrightharpoonup: "\u21c0",
+  longleftharpoondown: "\u21bd",
   "\\cdrightarrow": "\u2192",
   "\\cdleftarrow": "\u2190",
   "\\cdlongequal": "="
 };
 
-// For stretchy elements for which no font glyph exists.
-const svgData = {
-  longRightleftharpoons: { minWidth: 1.75,  height: 0.716 },
-  longLeftrightharpoons: { minWidth: 1.75,  height: 0.716 }
+const nodeFromObject = (obj) => {
+  // Build a stretchy arrow from two SVGs.
+  const children = [];
+  if (obj.children) {
+    obj.children.map(child => { children.push(nodeFromObject(child)); });
+  }
+  const node = obj.type === "span"
+    ? new Span(null, children, obj.style)
+    : new mathMLTree.MathNode(obj.type, children, [], obj.style, true);
+  Object.entries(obj.attributes).forEach(([key, value]) => {
+    node.setAttribute(key, value);
+  });
+  return node
 };
 
-const setSvgAttributes = (svg, key, aspect) => {
-  svg.setAttribute("xmlns", "http://www.w3.org/2000/svg");
-  svg.setAttribute("width", "400em");
-  svg.setAttribute("height", svgData[key].height + "em");
-  svg.setAttribute("viewBox", `0 0 400000 ${1000 * svgData[key].height}`);
-  svg.setAttribute("preserveAspectRatio", aspect);
-  svg.setAttribute("fill", "currentColor");
-  svg.setAttribute("fill-rule", "non-zero");
-  svg.setAttribute("fill-opacity", "1");
-  return svg
-};
-
-const stretchySVG = (key, macros) => {
-  // No Unicode glyph matches the desired arrow.
-  // So we'll create our own arrow from SVGs.
-
-  // Two SVGs, one for each end of the arrow. The SVGs are very long (400em).
-  const leftPath = new mathMLTree.MathNode("path", [], [], {}, true);
-  leftPath.setAttribute("stroke", "none");
-  const leftPathGeometry = macros.get(key + "Left");
-  leftPath.setAttribute("d", leftPathGeometry);
-
-  const rightPath = new mathMLTree.MathNode("path", [], [], {}, true);
-  rightPath.setAttribute("stroke", "none");
-  const rightPathGeometry = macros.get(key + "Right");
-  rightPath.setAttribute("d", rightPathGeometry);
-
-  let leftSVG = new mathMLTree.MathNode("svg", [leftPath], [], {}, true);
-  leftSVG = setSvgAttributes(leftSVG, key, "xMinYMin slice");
-  leftSVG.setAttribute("style", "position: absolute; left:0;");
-
-  let rightSVG = new mathMLTree.MathNode("svg", [rightPath], [], {}, true);
-  rightSVG = setSvgAttributes(rightSVG, key, "xMaxYMin slice");
-  rightSVG.setAttribute("style", "position: absolute; right:0");
-
-  const height = svgData[key].height;
-
-  // Wrap the SVG in a span. Set the span width to 50% and set overflow: hidden.
-  // That enables us to stretch the arrow without deforming the arrowhead.
-  const leftSpan = new Span([], [leftSVG]);
-  leftSpan.setAttribute(
-    "style",
-    `position: absolute; left: 0; width: 50.2%; height: ${height}em; overflow: hidden;`
-  );
-
-  const rightSpan = new Span([], [rightSVG]);
-  rightSpan.setAttribute(
-    "style",
-    `position: absolute; right: 0; width: 50.2%; height: ${height}em; overflow: hidden;`
-  );
-
-  const wrapper = new Span(["stretchy"], [leftSpan, rightSpan]);
-  wrapper.setAttribute(
-    "style",
-    `display: block; position: relative; width: 100%; height: ${height}em;
-min-width: ${svgData[key].minWidth}em;`
-  );
-
-  return wrapper
-};
-
-const mathMLnode = function(label, macros) {
+const mathMLnode = function(label, macros = {}) {
   const key = label.slice(1);
-  const child = utils.contains(keysWithoutUnicodePoints, key)
-    ? stretchySVG(key, macros)
-    : new mathMLTree.TextNode(stretchyCodePoint[key]);
+  let child;
+  if (!keysWithoutUnicodePoints.includes(key)) {
+    child = new mathMLTree.TextNode(stretchyCodePoint[key]);
+  } else {
+    const atKey = "\\@" + key;
+    if (!macros.has(atKey)) {
+      throw new ParseError("Arrow not available. The mhchem package is needed.")
+    }
+    child = nodeFromObject(JSON.parse(macros.get(atKey)));
+  }
   const node = new mathMLTree.MathNode("mo", [child]);
   node.setAttribute("stretchy", "true");
   return node
@@ -1285,7 +1237,6 @@ defineSymbol(math, bin, "\u22cf", "\\curlywedge", true);
 defineSymbol(math, bin, "\u22ce", "\\curlyvee", true);
 defineSymbol(math, bin, "\u229d", "\\circleddash", true);
 defineSymbol(math, bin, "\u229b", "\\circledast", true);
-defineSymbol(math, bin, "\u22c5", "\\centerdot");
 defineSymbol(math, bin, "\u22ba", "\\intercal", true);
 defineSymbol(math, bin, "\u22d2", "\\doublecap");
 defineSymbol(math, bin, "\u22d3", "\\doublecup");
@@ -1441,6 +1392,7 @@ defineSymbol(math, mathord, "\u03da", "\\Stigma", true);
 defineSymbol(math, mathord, "\u03db", "\\stigma", true);
 defineSymbol(math, bin, "\u2217", "\u2217", true);
 defineSymbol(math, bin, "+", "+");
+defineSymbol(math, bin, "*", "*");
 defineSymbol(math, bin, "\u2212", "-", true);
 defineSymbol(math, bin, "\u22c5", "\\cdot", true);
 defineSymbol(math, bin, "\u2218", "\\circ");
@@ -2116,8 +2068,8 @@ function buildMathML(tree, texExpression, style, settings) {
   }
 
   const math = settings.annotate
-  ? new mathMLTree.MathNode("math", [semantics])
-  : new mathMLTree.MathNode("math", [wrapper]);
+    ? new mathMLTree.MathNode("math", [semantics])
+    : new mathMLTree.MathNode("math", [wrapper]);
 
   if (settings.xml) {
     math.setAttribute("xmlns", "http://www.w3.org/1998/Math/MathML");
@@ -2285,12 +2237,36 @@ defineFunction({
   }
 });
 
-// Helper function
-const paddedNode$1 = (group, width = "+0.6em") => {
+// Helper functions
+const paddedNode = (group, width = "+0.6em") => {
   const node = new mathMLTree.MathNode("mpadded", group ? [group] : []);
   node.setAttribute("width", width);
   node.setAttribute("lspace", "0.3em");
   return node;
+};
+
+const munderoverNode = (label, body, below, style, macros = {}) => {
+  const arrowNode = stretchy.mathMLnode(label, macros);
+  const minWidth = label.charAt(1) === "x"
+    ? "1.75em"  // mathtools extensible arrows
+    : label.slice(2, 4) === "cd"
+    ? "3.0em"  // cd package arrows
+    : "2.0em"; // mhchem arrows
+  arrowNode.setAttribute("minsize", minWidth);
+  // minsize attribute doesn't work in Firefox.
+  // https://bugzilla.mozilla.org/show_bug.cgi?id=320303
+  const labelStyle = style.incrementLevel();
+
+  const upperNode = (body && body.body && body.body.length > 0)
+    ? paddedNode(buildGroup(body, labelStyle))
+      // Since Firefox does not recognize minsize set on the arrow,
+      // create an upper node w/correct width.
+    : paddedNode(null, minWidth);
+  const lowerNode = (below && below.body && below.body.length > 0)
+    ? paddedNode(buildGroup(below, labelStyle))
+    : paddedNode(null, minWidth);
+  const node = new mathMLTree.MathNode("munderover", [arrowNode, lowerNode, upperNode]);
+  return node
 };
 
 // Stretchy arrows with an optional argument
@@ -2310,18 +2286,17 @@ defineFunction({
     "\\xrightharpoonup",
     "\\xleftharpoondown",
     "\\xleftharpoonup",
-    "\\xrightleftharpoons",
-    "\\xleftrightharpoons",
     "\\xlongequal",
     "\\xtwoheadrightarrow",
     "\\xtwoheadleftarrow",
-    "\\xtofrom",
-    // The next 4 functions are here to support the mhchem extension.
-    // Direct use of these functions is discouraged and may break someday.
-    "\\longrightleftharpoons",
-    "\\longleftrightarrows",   // <--> or
-    "\\longRightleftharpoons",
-    "\\longLeftrightharpoons",
+    // The next 7 functions are here only to support mhchem
+    "\\yields",
+    "\\yieldsLeft",
+    "\\mesomerism",
+    "\\longrightharpoonup",
+    "\\longleftharpoondown",
+    "\\equilibriumRight",
+    "\\equilibriumLeft",
     // The next 3 functions are here only to support the {CD} environment.
     "\\\\cdrightarrow",
     "\\\\cdleftarrow",
@@ -2342,23 +2317,82 @@ defineFunction({
     };
   },
   mathmlBuilder(group, style) {
-    const arrowNode = stretchy.mathMLnode(group.label, group.macros);
-    const minWidth = group.label.charAt(1) === "x" ? "1.75em" : "3.0em";
-    arrowNode.setAttribute("minsize", minWidth);
-    // minsize attribute doesn't work in Firefox.
-    // https://bugzilla.mozilla.org/show_bug.cgi?id=320303
-    const labelOptions = style.incrementLevel();
+    return munderoverNode(group.label, group.body, group.below, style, group.macros)
+  }
+});
 
-    const upperNode = (group.body && group.body.body.length > 0)
-      ? paddedNode$1(buildGroup(group.body, labelOptions))
-        // Since Firefox does not recognize minsize set on the arrow,
-        // create an upper node w/correct width.
-      : paddedNode$1(null, minWidth);
-    const lowerNode = (group.below && group.below.body.length > 0)
-      ? paddedNode$1(buildGroup(group.below, labelOptions))
-      : paddedNode$1(null, minWidth);
-    const node = new mathMLTree.MathNode("munderover", [arrowNode, lowerNode, upperNode]);
-    return node;
+const arrowComponent = {
+  "\\xtofrom": ["\\xrightarrow", "\\xleftarrow"],
+  "\\xleftrightharpoons": ["\\xleftharpoonup", "\\xrightharpoondown"],
+  "\\xrightleftharpoons": ["\\xrightharpoonup", "\\xleftharpoondown"],
+  "\\yieldsLeftRight": ["\\yields", "\\yieldsLeft"],
+  "\\equilibrium": ["\\longrightharpoonup", "\\longleftharpoondown"]
+};
+
+// Browser are not good at stretching stacked arrows such as ⇄.
+// So we stack a pair of single arrows.
+defineFunction({
+  type: "stackedArrow",
+  names: [
+    "\\xtofrom",              // expfeil
+    "\\xleftrightharpoons",   // mathtools
+    "\\xrightleftharpoons",   // mathtools
+    "\\yieldsLeftRight",      // mhchem
+    "\\equilibrium"           // mhchem
+  ],
+  props: {
+    numArgs: 1,
+    numOptionalArgs: 1
+  },
+  handler({ parser, funcName }, args, optArgs) {
+    const lowerArrowBody = args[0]
+      ? {
+        type: "hphantom",
+        mode: parser.mode,
+        body: args[0]
+      }
+      : null;
+    const upperArrowBelow = optArgs[0]
+      ? {
+        type: "hphantom",
+        mode: parser.mode,
+        body: optArgs[0]
+      }
+      : null;
+    return {
+      type: "stackedArrow",
+      mode: parser.mode,
+      label: funcName,
+      body: args[0],
+      upperArrowBelow,
+      lowerArrowBody,
+      below: optArgs[0]
+    };
+  },
+  mathmlBuilder(group, style) {
+    const topLabel = arrowComponent[group.label][0];
+    const botLabel = arrowComponent[group.label][1];
+    const topArrow = munderoverNode(topLabel, group.body, group.upperArrowBelow, style);
+    const botArrow = munderoverNode(botLabel, group.lowerArrowBody, group.below, style);
+
+    const topSpace = new mathMLTree.MathNode("mspace");
+    topSpace.setAttribute("width", "0.2778em");
+    const botSpace = new mathMLTree.MathNode("mspace");
+    botSpace.setAttribute("width", "-0.2778em");
+
+    const raiseNode = new mathMLTree.MathNode("mpadded", [topSpace, topArrow]);
+    raiseNode.setAttribute("voffset", "0.3em");
+    raiseNode.setAttribute("height", "+0.3em");
+    raiseNode.setAttribute("depth", "-0.3em");
+    raiseNode.setAttribute("width", "0em");
+
+    const botRow = new mathMLTree.MathNode("mrow", [botSpace, botArrow]);
+
+    const wrapper = new mathMLTree.MathNode("mpadded", [raiseNode, botRow]);
+    wrapper.setAttribute("voffset", "-0.18em");
+    wrapper.setAttribute("height", "-0.18em");
+    wrapper.setAttribute("depth", "+0.18em");
+    return wrapper
   }
 });
 
@@ -2731,42 +2765,45 @@ const toHex = num => {
 };
 
 // Colors from Tables 4.1 and 4.2 of the xcolor package.
+// Table 4.1 (lower case) RGB values are taken from chroma and xcolor.dtx.
+// Table 4.2 (Capitalizzed) values were sampled, because Chroma contains a unreliable
+// conversion from cmyk to RGB. See https://tex.stackexchange.com/a/537274.
 const xcolors = JSON.parse(`{
   "Apricot": "#ffb484",
   "Aquamarine": "#08b4bc",
   "Bittersweet": "#c84c14",
-  "blue": "#0804fc",
+  "blue": "#0000FF",
   "Blue": "#303494",
   "BlueGreen": "#08b4bc",
   "BlueViolet": "#503c94",
   "BrickRed": "#b8341c",
-  "brown": "#c08444",
+  "brown": "#BF804",
   "Brown": "#802404",
   "BurntOrange": "#f8941c",
   "CadetBlue": "#78749c",
   "CarnationPink": "#f884b4",
   "Cerulean": "#08a4e4",
   "CornflowerBlue": "#40ace4",
-  "cyan": "#08acec",
+  "cyan": "#00FFF",
   "Cyan": "#08acec",
   "Dandelion": "#ffbc44",
-  "darkgray": "#484444",
+  "darkgray": "#404040",
   "DarkOrchid": "#a8548c",
   "Emerald": "#08ac9c",
   "ForestGreen": "#089c54",
   "Fuchsia": "#90348c",
   "Goldenrod": "#ffdc44",
-  "gray": "#888484",
+  "gray": "#80808",
   "Gray": "#98949c",
-  "green": "#08fc04",
+  "green": "#00FF0",
   "Green": "#08a44c",
   "GreenYellow": "#e0e474",
   "JungleGreen": "#08ac9c",
   "Lavender": "#f89cc4",
-  "lightgray": "#c0bcbc",
-  "lime": "#c0fc04",
+  "lightgray": "#BFBFB",
+  "lime": "#BFFF00",
   "LimeGreen": "#90c43c",
-  "magenta": "#f0048c",
+  "magenta": "#FF00F",
   "Magenta": "#f0048c",
   "Mahogany": "#b0341c",
   "Maroon": "#b03434",
@@ -2774,22 +2811,22 @@ const xcolors = JSON.parse(`{
   "MidnightBlue": "#086494",
   "Mulberry": "#b03c94",
   "NavyBlue": "#086cbc",
-  "olive": "#988c04",
+  "olive": "#7F7F00",
   "OliveGreen": "#407c34",
-  "orange": "#ff8404",
+  "orange": "#FF800",
   "Orange": "#f8843c",
   "OrangeRed": "#f0145c",
   "Orchid": "#b074ac",
   "Peach": "#f8945c",
   "Periwinkle": "#8074bc",
   "PineGreen": "#088c74",
-  "pink": "#ffbcbc",
+  "pink": "#ff7f7f",
   "Plum": "#98248c",
   "ProcessBlue": "#08b4ec",
-  "purple": "#c00444",
+  "purple": "#BF0040",
   "Purple": "#a0449c",
   "RawSienna": "#983c04",
-  "red": "#ff0404",
+  "red": "#ff0000",
   "Red": "#f01c24",
   "RedOrange": "#f86434",
   "RedViolet": "#a0246c",
@@ -2803,15 +2840,15 @@ const xcolors = JSON.parse(`{
   "SkyBlue": "#48c4dc",
   "SpringGreen": "#c8dc64",
   "Tan": "#e09c74",
-  "teal": "#088484",
+  "teal": "#007F7F",
   "TealBlue": "#08acb4",
   "Thistle": "#d884b4",
   "Turquoise": "#08b4cc",
-  "violet": "#880484",
+  "violet": "#800080",
   "Violet": "#60449c",
   "VioletRed": "#f054a4",
   "WildStrawberry": "#f0246c",
-  "yellow": "#fff404",
+  "yellow": "#FFFF00",
   "Yellow": "#fff404",
   "YellowGreen": "#98cc6c",
   "YellowOrange": "#ffa41c"
@@ -4046,7 +4083,11 @@ const mathmlBuilder$3 = function(group, style) {
   for (let i = 0; i < numRows; i++) {
     const rw = group.body[i];
     const row = [];
-    const cellStyle = group.scriptLevel === "text" ? StyleLevel.TEXT : StyleLevel.DISPLAY;
+    const cellStyle = group.scriptLevel === "text"
+    ? StyleLevel.TEXT
+    : group.scriptLevel === "script"
+    ? StyleLevel.SCRIPT
+    : StyleLevel.DISPLAY;
 
     for (let j = 0; j < rw.length; j++) {
       const mtd = new mathMLTree.MathNode(
@@ -4734,7 +4775,7 @@ function mathmlBuilder$4(group, style) {
   if (group.mclass === "minner") {
     return mathMLTree.newDocumentFragment(inner);
   } else if (group.mclass === "mord") {
-    if (group.isCharacterBox) {
+    if (group.isCharacterBox || inner[0].type === "mathord") {
       node = inner[0];
       node.type = "mi";
     } else {
@@ -4753,18 +4794,26 @@ function mathmlBuilder$4(group, style) {
 
     // Set spacing based on what is the most likely adjacent atom type.
     // See TeXbook p170.
+    const doSpacing = style.level < 2; // Operator spacing is zero inside a (sub|super)script.
     if (group.mclass === "mbin") {
-      node.attributes.lspace = "0.22em"; // medium space
-      node.attributes.rspace = "0.22em";
+      // medium space
+      node.attributes.lspace = (doSpacing ? "0.2222em" : "0");
+      node.attributes.rspace = (doSpacing ? "0.2222em" : "0");
+    } else if (group.mclass === "mrel") {
+      // thickspace
+      node.attributes.lspace = (doSpacing ? "0.2778em" : "0");
+      node.attributes.rspace = (doSpacing ? "0.2778em" : "0");
     } else if (group.mclass === "mpunct") {
       node.attributes.lspace = "0em";
-      node.attributes.rspace = "0.17em"; // thinspace
+      node.attributes.rspace = (doSpacing ? "0.1667em" : "0");
     } else if (group.mclass === "mopen" || group.mclass === "mclose") {
       node.attributes.lspace = "0em";
       node.attributes.rspace = "0em";
     }
-    // MathML <mo> default space is 5/18 em, so <mrel> needs no action.
-    // Ref: https://developer.mozilla.org/en-US/docs/Web/MathML/Element/mo
+    if (!(group.mclass === "mopen" || group.mclass === "mclose")) {
+      delete node.attributes.stretchy;
+      delete node.attributes.form;
+    }
   }
   return node;
 }
@@ -4790,32 +4839,27 @@ defineFunction({
     // We should not wrap a <mo> around a <mi> or <mord>. That would be invalid MathML.
     // In that case, we instead promote the text contents of the body to the parent.
     let mustPromote = true;
-    const atom = { type: "atom", family: funcName.substr(5), text:"" };
+    const mord = { type: "mathord", text: "", mode: parser.mode };
     const arr = (body.body) ? body.body : [body];
     for (const arg of arr) {
       if (utils.textAtomTypes.includes(arg.type)) {
         if (arg.text) {
-          atom.text += arg.text;
+          mord.text += arg.text;
         } else if (arg.body) {
-          arg.body.map(e => { atom.text += e.text; });
+          arg.body.map(e => { mord.text += e.text; });
         }
       } else {
         mustPromote = false;
         break
       }
     }
-    if (mustPromote) {
-      atom.mode = parser.mode;
-      return atom
-    } else {
-      return {
-        type: "mclass",
-        mode: parser.mode,
-        mclass: "m" + funcName.substr(5),
-        body: ordargument(body),
-        isCharacterBox: utils.isCharacterBox(body)
-      };
-    }
+    return {
+      type: "mclass",
+      mode: parser.mode,
+      mclass: "m" + funcName.substr(5),
+      body: ordargument(mustPromote ? mord : body),
+      isCharacterBox: utils.isCharacterBox(body) || mustPromote
+    };
   },
   mathmlBuilder: mathmlBuilder$4
 });
@@ -5828,7 +5872,7 @@ defineFunction({
 
 // Switching from text mode back to math mode
 defineFunction({
-  type: "styling",
+  type: "ordgroup",
   names: ["\\(", "$"],
   props: {
     numArgs: 0,
@@ -5843,9 +5887,8 @@ defineFunction({
     parser.expect(close);
     parser.switchMode(outerMode);
     return {
-      type: "styling",
+      type: "ordgroup",
       mode: parser.mode,
-      scriptLevel: "text",
       body
     };
   }
@@ -5987,50 +6030,6 @@ defineFunction({
     }
 
     return new mathMLTree.MathNode("mmultiscripts", children);
-  }
-});
-
-// The \newextarrow macro expands to this function.
-defineFunction({
-  type: "newExtArrow",
-  names: ["\\ext@arrow"],
-  props: {
-    numArgs: 4,
-    argTypes: ["raw", "raw", "raw", "math"]
-  },
-  handler({ parser }, args) {
-    return {
-      type: "newExtArrow",
-      mode: parser.mode,
-      lspace: utils.round(Number(args[0].string.trim()) / 18),
-      rspace: utils.round(Number(args[1].string.trim()) / 18),
-      charCode: Number(args[2].string.trim()),
-      body: args[3]
-    };
-  },
-  mathmlBuilder(group, style) {
-    const arrowNode = new mathMLTree.MathNode(
-      "mo",
-      [new mathMLTree.TextNode(String.fromCodePoint(group.charCode))]
-    );
-    arrowNode.setAttribute("stretchy", "true");
-    arrowNode.setAttribute("minsize", "1.75em");
-
-    let node;
-    if (group.body) {
-      let upperNode = buildGroup(group.body, style.incrementLevel());
-      upperNode = new mathMLTree.MathNode("mpadded", [upperNode]);
-      upperNode.setAttribute("width", "+" + (group.lspace + group.rspace) + "em");
-      upperNode.setAttribute("lspace", group.lspace + "em");
-      node = new mathMLTree.MathNode("mover", [arrowNode, upperNode]);
-    } else {
-      // This should never happen.
-      // Parser.js throws an error if there is no argument.
-      // eslint-disable-next-line no-undef
-      node = paddedNode();
-      node = new mathMLTree.MathNode("mover", [arrowNode, node]);
-    }
-    return node;
   }
 });
 
@@ -6769,12 +6768,12 @@ defineFunction({
     numArgs: 0,
     allowedInText: true
   },
-  handler({parser}) {
+  handler({ parser }) {
     return {
       type: "internal",
       mode: parser.mode
     };
-  },
+  }
 });
 
 defineFunction({
@@ -7028,6 +7027,9 @@ defineFunction({
  * handling them itself.
  */
 
+// Helpers
+const symbolRegEx = /^m(over|under|underover)$/;
+
 // Super scripts and subscripts, whose precise placement can depend on other
 // functions that precede them.
 defineFunctionBuilders({
@@ -7059,13 +7061,13 @@ defineFunctionBuilders({
       ? [buildGroup(group.base.body[0], style)]
       : [buildGroup(group.base, style)];
 
-    const childOptions = style.incrementLevel();
+    const childStyle = style.inSubOrSup();
     if (group.sub) {
-      children.push(buildGroup(group.sub, childOptions));
+      children.push(buildGroup(group.sub, childStyle));
     }
 
     if (group.sup) {
-      children.push(buildGroup(group.sup, childOptions));
+      children.push(buildGroup(group.sup, childStyle));
     }
 
     let nodeType;
@@ -7139,6 +7141,9 @@ defineFunctionBuilders({
       } else {
         node = mathMLTree.newDocumentFragment([node, operator]);
       }
+    } else if (symbolRegEx.test(nodeType)) {
+      // Wrap in a <mrow>. Otherwise Firefox stretchy parens will not stretch to include limits.
+      node = new mathMLTree.MathNode("mrow", [node]);
     }
 
     return node
@@ -7651,7 +7656,7 @@ defineFunction({
     };
   },
   mathmlBuilder(group, style) {
-    const newStyle = styleWithFont(group, style.withLevel(StyleLevel.TEXT));
+    const newStyle = styleWithFont(group, style);
     const mrow = buildExpressionRow(group.body, newStyle);
 
     // If possible, consolidate adjacent <mtext> elements into a single element.
@@ -7778,33 +7783,6 @@ defineFunction({
     node.setAttribute("accentunder", "true");
 
     return node;
-  }
-});
-
-// \vcenter:  Vertically center the argument group on the math axis.
-
-defineFunction({
-  type: "vcenter",
-  names: ["\\vcenter"],
-  props: {
-    numArgs: 1,
-    argTypes: ["math"],
-    allowedInText: false
-  },
-  handler({ parser }, args) {
-    return {
-      type: "vcenter",
-      mode: parser.mode,
-      body: args[0]
-    };
-  },
-  mathmlBuilder(group, style) {
-    // There is no way to do this in MathML.
-    // Write a class so a post-processor can easily find this element.
-    // TODO: Monitor MathML and update this when possible.
-    return new mathMLTree.MathNode("mpadded",
-      [buildGroup(group.body, style)], ["vcenter"]
-    );
   }
 });
 
@@ -8779,26 +8757,6 @@ defineMacro("\\pdv@next", function(context) {
 });
 
 //////////////////////////////////////////////////////////////////////
-// extpfeil package
-// \newextarrow{\arrowName}{lspace,rspace}{unicode-charcode}  lspace & rspace are in mu.
-// Unlike the extpfeil package, this command does not support an optional lower note.
-defineMacro("\\newextarrow", function(context) {
-  const args = context.consumeArgs(3);
-  const arrowName = args[0][0].text;
-  const padText = stringFromArg(args[1]);
-  const padNums = padText.split(",");
-  if (padNums.length !== 2 || isNaN(padNums[0] || isNaN(padNums[1]))) {
-    throw new ParseError("Invalid lspace,rspace in \\newextarrow.");
-  }
-  const charText = stringFromArg(args[2]);
-  if (isNaN(charText)) {
-    throw new ParseError("Invalid Unicode character code in \\newextarrow.");
-  }
-  defineMacro(arrowName, `\\ext@arrow{${padNums[0]}}{${padNums[1]}}{${charText}}{#1}`);
-  return "";
-});
-
-//////////////////////////////////////////////////////////////////////
 // upgreek.dtx
 defineMacro("\\upalpha", "\\up@greek{\\alpha}");
 defineMacro("\\upbeta", "\\up@greek{\\beta}");
@@ -8845,7 +8803,7 @@ defineMacro("\\standardstate", "{\\tiny\\char`⦵}");
  *    2. \rlap and \llap are replaced with \mathrlap and \mathllap.
  *    3. The reaction arrow code is simplified. All reaction arrows are rendered
  *       using Temml extensible arrows instead of building non-extensible arrows.
- *    4. SVG path geometry is supplied for \longRightleftharpoons & \longLeftrightharpoons.
+ *    4. SVG path geometry is supplied for \equilibriumRight & \equilibriumLeft.
  *    5. \tripledash uses Unicode character U+2504, ┄.
  *    6. Two dashes in _getBond are wrapped in braces to suppress spacing. i.e., {-}
  *    7. The electron dot uses \textbullet instead of \bullet.
@@ -8922,6 +8880,82 @@ c4.7,-4.7,7,-9.3,7,-14c0,-9.3,-3.7,-15.3,-11,-18c-92.7,-56.7,-159,-133.7,-199,
 -231c-3.3,-9.3,-6,-14.7,-8,-16c-2,-1.3,-7,-2,-15,-2c-10.7,0,-16.7,2,-18,6
 c-2,2.7,-1,9.7,3,21c15.3,42,36.7,81.8,64,119.5c27.3,37.7,58,69.2,92,94.5z
 M500 241 v40 H399408 v-40z M500 435 v40 H400000 v-40z`)
+
+
+const path = {
+  "\\equilibriumRight" : {
+    left: `M507,435c-4,4,-6.3,8.7,-7,14
+c0,5.3,0.7,9,2,11c1.3,2,5.3,5.3,12,10c90.7,54,156,130,196,228c3.3,10.7,6.3,16.3,9,17
+c2,0.7,5,1,9,1c0,0,5,0,5,0c10.7,0,16.7,-2,18,-6c2,-2.7,1,-9.7,-3,-21
+c-32,-87.3,-82.7,-157.7,-152,-211c0,0,-3,-3,-3,-3l399351,0l0,-40
+c-398570,0,-399437,0,-399437,0z M593 435 v40 H399500 v-40z
+M0 281 v-40 H399908 v40z M0 281 v-40 H399908 v40z`,
+    right: `M0,241 l0,40c399126,0,399993,0,399993,0
+c4.7,-4.7,7,-9.3,7,-14c0,-9.3,-3.7,-15.3,-11,-18c-92.7,-56.7,-159,-133.7,-199,
+-231c-3.3,-9.3,-6,-14.7,-8,-16c-2,-1.3,-7,-2,-15,-2c-10.7,0,-16.7,2,-18,6
+c-2,2.7,-1,9.7,3,21c15.3,42,36.7,81.8,64,119.5c27.3,37.7,58,69.2,92,94.5z
+M0 241 v40 H399908 v-40z M0 475 v-40 H399500 v40z M0 475 v-40 H399500 v40z`
+  },
+  "\\equilibriumLeft": {
+    left: `M7,435c-4,4,-6.3,8.7,-7,14c0,5.3,0.7,9,2,11
+c1.3,2,5.3,5.3,12,10c90.7,54,156,130,196,228c3.3,10.7,6.3,16.3,9,17c2,0.7,5,1,9,
+1c0,0,5,0,5,0c10.7,0,16.7,-2,18,-6c2,-2.7,1,-9.7,-3,-21c-32,-87.3,-82.7,-157.7,
+-152,-211c0,0,-3,-3,-3,-3l399907,0l0,-40c-399126,0,-399993,0,-399993,0z
+M93 435 v40 H400000 v-40z M500 241 v40 H400000 v-40z M500 241 v40 H400000 v-40z`,
+    right: `M53,241l0,40c398570,0,399437,0,399437,0
+c4.7,-4.7,7,-9.3,7,-14c0,-9.3,-3.7,-15.3,-11,-18c-92.7,-56.7,-159,-133.7,-199,
+-231c-3.3,-9.3,-6,-14.7,-8,-16c-2,-1.3,-7,-2,-15,-2c-10.7,0,-16.7,2,-18,6
+c-2,2.7,-1,9.7,3,21c15.3,42,36.7,81.8,64,119.5c27.3,37.7,58,69.2,92,94.5z
+M500 241 v40 H399408 v-40z M500 435 v40 H400000 v-40z`
+  }
+}
+
+const svg = (name, end) => {
+  const svg = {
+    type: "svg",
+    attributes: {
+      xmlns: "http://www.w3.org/2000/svg",
+      width: "400em",
+      height: "0.716em",
+      viewBox: "0 0 400000 716",
+      preserveAspectRatio: `x${end === "left" ? "Min" : "Max"}yMin slice`,
+      fill: "currrentColor",
+      "fill-rule": "non-zero",
+      "fill-opacity" : 1
+    },
+    style: { position: "absolute" },
+    children: [{
+      type: "path",
+      attributes: { d: path[name][end], stroke: "none" },
+      children: []
+    }]
+  }
+  if (end === "left") { svg.style.left = 0 } else { svg.style.right = 0 }
+  return svg
+}
+
+const endSpan = (name, end) => {
+  const span = {
+    type: "span",
+    attributes: {},
+    style: { position: "absolute", width: "50.2%", height: "0.716em", overflow: "hidden" },
+    children: [svg(name, end)]
+  }
+  if (end === "left") { span.style.left = 0 } else { span.style.right = 0 }
+  return span
+}
+
+const arrowNode = name => {
+  return JSON.stringify({
+    type: "span",
+    attributes: {},
+    style: { display: "block", position: "relative", width: "100%", height: "0.716em", "min-width": "1.75em" },
+    children: [endSpan(name, "left"), endSpan(name, "right")]
+  })
+}
+
+defineMacro("\\@equilibriumRight", arrowNode("\\equilibriumRight"))
+defineMacro("\\@equilibriumLeft", arrowNode("\\equilibriumLeft"))
 
   //
   //  This is the main function for handing the \ce and \pu commands.
@@ -10393,9 +10427,9 @@ M500 241 v40 H399408 v-40z M500 435 v40 H400000 v-40z`)
             rq: texify._goInner(buf.rq)
           };
           var arrow = texify._getArrow(buf.r);
-          if (b6.rq) { arrow += "[{" + b6.rq + "}]"; }
+          if (b6.rq) { arrow += "[{\\rm " + b6.rq + "}]"; }
           if (b6.rd) {
-            arrow += "{" + b6.rd + "}";
+            arrow += "{\\rm " + b6.rd + "}";
           } else {
             arrow += "{}";
           }
@@ -10485,16 +10519,16 @@ M500 241 v40 H399408 v-40z M500 435 v40 H400000 v-40z`)
     },
     _getArrow: function (a) {
       switch (a) {
-        case "->": return "\\xrightarrow";
-        case "\u2192": return "\\xrightarrow";
-        case "\u27F6": return "\\xrightarrow";
-        case "<-": return "\\xleftarrow";
-        case "<->": return "\\xleftrightarrow";
-        case "<-->": return "\\longleftrightarrows";
-        case "<=>": return "\\longrightleftharpoons";
-        case "\u21CC": return "\\longrightleftharpoons";
-        case "<=>>": return "\\longRightleftharpoons";
-        case "<<=>": return "\\longLeftrightharpoons";
+        case "->": return "\\yields";
+        case "\u2192": return "\\yields";
+        case "\u27F6": return "\\yields";
+        case "<-": return "\\yieldsLeft";
+        case "<->": return "\\mesomerism";
+        case "<-->": return "\\yieldsLeftRight";
+        case "<=>": return "\\equilibrium";
+        case "\u21CC": return "\\equilibrium";
+        case "<=>>": return "\\equilibriumRight";
+        case "<<=>": return "\\equilibriumLeft";
         default:
           assertNever(a);
           throw ["MhchemBugT", "mhchem bug T. Please report."];
@@ -12583,6 +12617,8 @@ const parseTree = function(toParse, settings) {
  * recursing, a new `Style` object can be created with the `.with*` functions.
  */
 
+const subOrSupLevel = [2, 2, 3, 3];
+
 /**
  * This is the main Style class. It contains the current style.level, color, and font.
  *
@@ -12642,6 +12678,12 @@ class Style {
     return this.extend({
       level: Math.min(this.level + 1, 3)
     });
+  }
+
+  inSubOrSup() {
+    return this.extend({
+      level: subOrSupLevel[this.level]
+    })
   }
 
   /**
@@ -12711,7 +12753,7 @@ class Style {
  * https://mit-license.org/
  */
 
-const version = "0.3.3";
+const version = "0.4.0";
 
 function postProcess(block) {
   const labelMap = {};